cleanup and add to READMEs
