+++ README.md
@@ -1,3 +1,19 @@
 # PowerSync Web SDK
 
-This repo contains the packages for PowerSync's Web SDK.
+This mono repo contains the packages for PowerSync's Web SDK.
+
+## Structure
+
+### Demos
+
+Demo applications are located in the `/demos/` directory.
+
+ - [apps/powersync-nextjs-demo](./packages/powersync-nextjs-demo/README.md): A NextJS 14 Todo list application.
+
+
+### Packages
+
+SDK packages are located in the `/packages/` directory.
+
+ - [packages/powersync-sdk-web](./packages/powersync-sdk-web/README.md) contains the core PowerSync Web SDK code.
+
