+++ demos/powersync-nextjs-demo/.env.local.template
