+++ demos/powersync-nextjs-demo/.eslintrc.json
