+++ demos/powersync-nextjs-demo/.gitignore
