+++ demos/powersync-nextjs-demo/next.config.js
@@ -5,7 +5,7 @@ const nextConfig = {
     config.resolve.alias = {
       ...config.resolve.alias,
       // TODO fix this
-      'wa-sqlite/src/types': 'react'
+      'wa-sqlite/src/types': '@journeyapps/powersync-sdk-web'
     };
     return config;
   }
