+++ demos/powersync-nextjs-demo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "powersync-test",
+  "name": "powersync-nextjs-demo",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -36,7 +36,6 @@
     "eslint": "^8",
     "eslint-config-next": "14.0.0",
     "postcss": "^8",
-    "tailwindcss": "^3",
     "typescript": "^5"
   }
 }
