+++ demos/powersync-nextjs-demo/public/supabase-logo.png
