+++ demos/powersync-nextjs-demo/src/app/api/supabase/route.ts
