+++ demos/powersync-nextjs-demo/src/app/favicon.ico
