+++ demos/powersync-nextjs-demo/src/app/globals.css
@@ -1,7 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 :root {
   --foreground-rgb: 0, 0, 0;
   --background-start-rgb: 214, 219, 220;
@@ -19,5 +15,6 @@
 body {
   color: rgb(var(--foreground-rgb));
   min-height: 100vh;
+  margin: 0;
   background: linear-gradient(to bottom, transparent, rgb(var(--background-end-rgb))) rgb(var(--background-start-rgb));
 }
