+++ demos/powersync-nextjs-demo/src/app/layout.tsx
