+++ demos/powersync-nextjs-demo/src/app/page.tsx
