+++ demos/powersync-nextjs-demo/src/app/views/layout.tsx
