+++ demos/powersync-nextjs-demo/src/components/DynamicParentProvider.tsx
