+++ demos/powersync-nextjs-demo/src/components/Routes.ts
