+++ demos/powersync-nextjs-demo/src/components/Theme.tsx
