+++ demos/powersync-nextjs-demo/src/components/navigation/NavigationPage.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { useNavigationPanel } from './NavigationPanelContext';
+import { Box, styled } from '@mui/material';
 
 /**
  * Wraps a component with automatic navigation panel title management
@@ -13,5 +14,11 @@ export const NavigationPage: React.FC<React.PropsWithChildren<{ title: string }>
     return () => navigationPanel.setTitle('');
   }, [title]);
 
-  return children;
+  return <S.Container>{children}</S.Container>;
 };
+
+namespace S {
+  export const Container = styled(Box)`
+    margin: 10px;
+  `;
+}
