+++ demos/powersync-nextjs-demo/src/components/navigation/NavigationPanelContext.tsx
