+++ demos/powersync-nextjs-demo/src/components/widgets/LoginDetailsWidget.tsx
