+++ demos/powersync-nextjs-demo/src/library/powersync/AppSchema.ts
