+++ demos/powersync-nextjs-demo/src/library/powersync/SupabaseConnector.ts
