+++ demos/powersync-nextjs-demo/src/pages/about.tsx
