+++ demos/powersync-nextjs-demo/tsconfig.json
