+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
   # all packages in direct subdirs of packages/
   - "packages/*"
-  - "apps/*"
+  # all demos in direct subdirs of demos/
+  - "demos/*"
